Move duration argparse helper to common
